Added Model Tests
