+++ realwork/Podfile
@@ -7,9 +7,8 @@ target 'realwork' do
 
   # Pods for realwork
 #  pod 'NLibrary', :git => 'https://github.com/NathanDVT/NLibrary.git', :branch => 'development'
-  pod 'NLibrary', :git => 'https://github.com/NathanDVT/NLibrary.git', :branch => 'feature/unit_test'
-
-#  pod 'NLibrary', :path => '/Users/nngobale/Documents/Training 2020/Pod_Folder/NLibrary/NLibrary'
+#  pod 'NLibrary', :git => 'https://github.com/NathanDVT/NLibrary.git', :branch => 'feature/unit_test'
+  pod 'NLibrary', :path => '/Users/nngobale/Documents/Training 2020/Pod_Folder/NLibrary/NLibrary'
 
   target 'realworkTests' do
     inherit! :search_paths
